Calibrations

Here we implement functions for deriving the absolute magnitude of the TRGB from various literature sources.

TRGBDistances.Rizzi2007 — Function

Rizzi2007(color, obssym::Symbol, magsym::Symbol, colorsym::Symbol)

Color-dependent TRGB zeropoint magnitude from Rizzi et al. [6].

Supported filter combinations are:

obssym	magsym	colorsym	Equation

:JC	:I	:VI	-4.05 + 0.217 * (color - 1.60)

:ACS	:F814W	:F606W	-4.06 + 0.200 * (color - 1.23)

:ACS	:F814W	:F555W	-4.06 + 0.150 * (color - 1.74)

:WFPC2	:F814W	:F555W	-4.01 + 0.180 * (color - 1.58)

:WFPC2	:F814W	:F606W	-4.01 + 0.150 * (color - 1.12)

Returns

Float64: The TRGB zeropoint magnitude in the requested band.

Examples

julia> TRGBDistances.Rizzi2007(1.5, :ACS, :F814W, :F606W) ≈ -4.006
true

TRGBDistances.Fusco2012 — Function

Fusco2012(color)

Color-dependent TRGB zeropoint magnitude from Fusco et al. [8], $M_{F814W} = -3.63 + 0.08 * (F475W - F814W)^2 - 0.4 * (F475W - F814W)$.

julia> TRGBDistances.Fusco2012(1.5) ≈ -4.05
true

TRGBDistances.Jang2017 — Function

Jang2017(color, obssym::Symbol, magsym::Symbol, colorsym::Symbol)

Color-dependent TRGB zeropoint magnitudes from Jang and Lee [9]. Supported filter combinations are:

obssym	magsym	colorsym	Equation

:JC	:I	:V	-4.016 + 0.091*(color - 1.5)^2 - 0.007*(color - 1.5)

:ACS	:F814W	:F606W	-4.015 + 0.159*(color - 1.1)^2 - 0.047*(color - 1.1)

:ACS	:F814W	:F555W	-4.011 + 0.116*(color - 1.6)^2 - 0.043*(color - 1.6)

:WFC3	:F814W	:F606W	-4.024 + 0.150*(color - 1.1)^2 - 0.050*(color - 1.1)

:WFC3	:F814W	:F555W	-4.019 + 0.113*(color - 1.6)^2 - 0.048*(color - 1.6)

:WFPC2	:F814W	:F555W	-4.007 + 0.123*(color - 1.5)^2 - 0.051*(color - 1.5)

:WFPC2	:F814W	:F606W	-3.984 + 0.179*(color - 1.0)^2 - 0.034*(color - 1.0)

julia> TRGBDistances.Jang2017(1.5, :ACS, :F814W, :F606W) ≈ -4.00836
true

TRGBDistances.Freedman2019 — Function

Freedman2019()

TRGB zeropoint from Freedman et al. [10] of -4.049 ± 0.022 (stat) ± 0.039 (sys) in I and HST/F814W.

julia> TRGBDistances.Freedman2019() ≈ -4.049
true

TRGBDistances.Freedman2020 — Function

Freedman2020(color::Union{Real, Nothing}=nothing,
             magsym::Symbol=:I)

TRGB zeropoint calibration from Freedman et al. [11].

Supported parameter combinations are:

magsym	Assumed color convention	Equation

:I	none	-4.047

:F814W	none	-4.054

:J	J - K	-5.14 - 0.85 * (color - 1)

:H	J - K	-5.95 - 1.62 * (color - 1)

:K	J - K	-6.14 - 1.85 * (color - 1)

:V	V - I	-2.25 + (color - 1.8)

Arguments

color: Optional color term used for color-dependent calibrations. For :J, :H, and :K, the assumed color is J - K. For :V, the assumed color is V - I.

magsym: Magnitude/filter band in which the TRGB zeropoint is returned.

Returns

Float64: The TRGB zeropoint magnitude in the requested band.

Examples

julia> TRGBDistances.Freedman2020(nothing, :I) ≈ -4.047
true

julia> TRGBDistances.Freedman2020(1.2, :J) ≈ -5.31
true

TRGBDistances.Yuan2019 — Function

Yuan2019(obssym::Symbol=:ACS, magsym::Symbol=:F814W)

TRGB zeropoint from Yuan et al. [12], which is -3.97 ± 0.046 in F814W and I.

julia> TRGBDistances.Yuan2019() ≈ -3.97
true

TRGBDistances.Cerny2021 — Function

Cerny2021(color::Union{Real, Nothing}=nothing,
           magsym::Symbol=:I,
           colorsym::Union{Symbol, Nothing}=nothing)

TRGB zeropoint calibration from Cerny et al. [13].

Supported parameter combinations are:

magsym	colorsym	Equation

:I	nothing	-4.056

:F814W	nothing	-4.056

:V	nothing	-2.26

:J	nothing	-5.16

:H	nothing	-6.03

:K	nothing	-6.16

:I	:VI	-4.06

:F814W	:VI	-4.06

:V	:VI	-2.26 + (color - 1.8)

:J	:JK	-5.16 - 0.85 * (color - 1)

:H	:JK	-6.03 - 1.72 * (color - 1)

:K	:JK	-6.16 - 1.85 * (color - 1)

Arguments

color: Optional color term used for color-dependent calibrations.

magsym: Magnitude/filter band in which the TRGB zeropoint is returned.

colorsym: Symbol describing the color combination associated with color.

Returns

Float64: The TRGB zeropoint magnitude in the requested band.

Examples

julia> TRGBDistances.Cerny2021() ≈ -4.056
true

julia> TRGBDistances.Cerny2021(1.2, :J, :JK) ≈ -5.33
true

TRGBDistances.Hoyt2018 — Function

Hoyt2018(color::Real, magsym::Symbol=:K, colorsym::Symbol=:JK)

TRGB zeropoint from Hoyt et al. [14]. Supported parameter combinations are:

magsym	colorsym	Assumed color	Equation

:J	:JK	J - K	-5.14 - 0.85 * (color - 1)

:H	:JK	J - K	-5.94 - 1.62 * (color - 1)

:K	:JK	J - K	-6.14 - 1.85 * (color - 1)

:J	:JH	J - H	-5.13 - 1.11 * (color - 0.8)

:H	:JH	J - H	-5.93 - 2.11 * (color - 0.8)

:K	:JH	J - H	-6.13 - 2.41 * (color - 0.8)

Arguments

color: Color term used in the calibration relation. The interpretation depends on colorsym.

magsym: Magnitude/filter band in which the TRGB zeropoint is returned.

colorsym: Symbol specifying the color convention associated with color.

Returns

Float64: The TRGB zeropoint magnitude in the requested band.

julia> TRGBDistances.Hoyt2018(1.5, :K) ≈ -6.14 - 1.85 * (1.5 - 1.0)
true

References

This page cites the following references:

[6]

L. Rizzi, R. B. Tully, D. Makarov, L. Makarova, A. E. Dolphin, S. Sakai and E. J. Shaya. Tip of the Red Giant Branch Distances. II. Zero-Point Calibration. ApJ 661, 815–829 (2007).

[8]

F. Fusco, R. Buonanno, G. Bono, S. Cassisi, M. Monelli and A. Pietrinferni. Distance and reddening of the Local Group dwarf irregular galaxy NGC 6822. A & A 548, A129 (2012).

[9]

I. S. Jang and M. G. Lee. The Tip of the Red Giant Branch Distances to Type Ia Supernova Host Galaxies. IV. Color Dependence and Zero-point Calibration. ApJ 835, 28 (2017).

[10]

W. L. Freedman, B. F. Madore, D. Hatt, T. J. Hoyt, I. S. Jang, R. L. Beaton, C. R. Burns, M. G. Lee, A. J. Monson, J. R. Neeley, M. Phillips, J. Rich and M. Seibert. The Carnegie-Chicago Hubble Program. VIII. An Independent Determination of the Hubble Constant Based on the Tip of the Red Giant Branch. ApJ 882, 34 (2019).

[11]

W. L. Freedman, B. F. Madore, T. Hoyt, I. S. Jang, R. Beaton, M. G. Lee, A. Monson, J. Neeley and J. Rich. Calibration of the Tip of the Red Giant Branch (TRGB). ApJ 891, 57 (2020).

[12]

W. Yuan, A. G. Riess, L. M. Macri, S. Casertano and D. Scolnic. Consistent Calibration of the Tip of the Red Giant Branch in the Large Magellanic Cloud on the Hubble Space Telescope Photometric System and a Redetermination of the Hubble Constant. ApJ 886, 61 (2019).

[13]

W. Cerny, W. L. Freedman, B. F. Madore, F. Ashmead, T. Hoyt, E. Oakes, N. Q. Tran and B. Moss. Multi-Wavelength, Optical (VI) and Near-Infrared (JHK) Calibration of the Tip of the Red Giant Branch Method based on Milky Way Globular Clusters. ArXiv e-prints (2020), arXiv:2012.09701 [astro-ph.GA].

[14]

T. J. Hoyt, W. L. Freedman, B. F. Madore, M. Seibert, R. L. Beaton, G. Clementini, I. S. Jang, M. G. Lee, A. Monson and J. Rich. The Near-infrared Tip of the Red Giant Branch. II. An Absolute Calibration in the Large Magellanic Cloud. ApJ 858, 12 (2018).