TRGBDistances

Documentation for TRGBDistances.

TRGBDistances.BrokenPowerLaw

TRGBDistances.LuminosityFunction

TRGBDistances.LuminosityFunctionData

TRGBDistances.Fusco2012

TRGBDistances.Rizzi2007

TRGBDistances.filter_mags

TRGBDistances.ψ

TRGBDistances.BrokenPowerLaw — Type

BrokenPowerLaw(m_trgb, a, b, c) <: LuminosityFunction

Standard broken power-law luminosity function typically used to model the luminosity function of the TRGB. For a magnitude m, the luminosity function is exp10(a * (m - m_trgb) + b) for m >= m_trgb and exp10(c * (m - m_trgb)) for m < m_trgb. Therefore a is the power-law slope of the RGB, b is the RGB bump, and c is the power-law slope of the AGB region.

For initial optimization conditions, eyeballing an approximate m_trgb on the CMD is usually sufficient. a = 0.3, b = 0.2, and c = 0.1 may be used as first guesses, but should be revised to match your data more closely.

TRGBDistances.LuminosityFunction — Type

LuminosityFunction{T} is the abstract supertype for luminosity functions that can be used to model the TRGB. Subtypes should implement ψ.

TRGBDistances.LuminosityFunctionData — Type

LuminosityFunctionData(model, mags,
                       err_func, complete_func, bias_func)

Structure to hold data for optimizing the given model. mags should typically be an array of apparent magnitudes with reddening correction applied. err_func(m) must return the photometric error at magnitude m, while complete_func(m) and bias_func(m) return the photometric completeness and bias (defined as <measured> - <input>), respectively.

When called with a vector of parameters θ, returns the negative logarithm of the likelihood (Equation 7 in Makarov+2006). This enables convenient optimization calls as most optimization routines will minimize an objective. For example, to optimize with the Nelder-Mead algorithm of Optim.jl, call result = Optim.optimize(LuminosityFunctionData(...), x0, Optim.NelderMead()). The maximum likelihood values are then Optim.minimizer(result).

TRGBDistances.Fusco2012 — Method

Fusco2012(color)

Returns M_{F814W} given a F475W - F814W color.

TRGBDistances.Rizzi2007 — Method

Rizzi2007(color, obssym::Symbol, magsym::Symbol, colorsym::Symbol)

Color-dependent TRGB zeropoint magnitude from Rizzi+2007.

julia> TRGBDistances.Rizzi2007(1.5, :ACS, :F814W, :F606W) ≈ -4.006
true

TRGBDistances.filter_mags — Method

filter_mags(colors, mags, ridge_colors, ridge_mags, func)

Returns indices into colors and mags of stars that fall within range of a ridgeline defined by ridge_colors and ridge_mags. The selection region is defined by the callable func(m) for magnitude m which must return a width. A star is accepted if its color falls between ± func(m) of the ridgeline.

TRGBDistances.ψ — Function

ψ(model::LuminosityFunction, m)

Returns the luminosity function of the model at magnitude m.